+++ src/main/java/com/waldronprojects/bookstore/entity/Product.java
@@ -12,7 +12,7 @@ public class Product {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(name="id")
-	private int id;
+	private Long id;
 	
 	@Column(name="product_name")
 	private String productName;
@@ -50,7 +50,7 @@ public class Product {
 		this.productTypeCollection = productTypeCollection;
 	}
 
-	public Product(int id, String productName, BigDecimal unitPrice, String description, int unitsInStock,
+	public Product(Long id, String productName, BigDecimal unitPrice, String description, int unitsInStock,
 				   Collection<ProductType> productTypeCollection) {
 		this.id = id;
 		this.productName = productName;
@@ -60,11 +60,11 @@ public class Product {
 		this.productTypeCollection = productTypeCollection;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
